Non multiples of four work for enemeis fireball wise

+++ Level.cpp
@@ -253,7 +253,7 @@ void Level::bossBattle()
         displayInput();
         
         // Check if correct number of fireballs killed
-        if (spritesDefeated == totalSpritesToKill * .25)
+        if (spritesDefeated == (int)(totalSpritesToKill * .25))
         {
             bossEndFightBegun = 1;
         }
